fix stats middleware
